修复路径删除/定位问题；优化区域完成时的填充效率；添加更新提醒；

+++ src/views/default-editor/area-viewer.vue
@@ -114,9 +114,13 @@
   onDeletePathwayEvent(() => {
     const currentPathway = controller.getCurrentPathway();
     if (props.layer && currentPathway) {
-      currentPathway.destroy();
-      controller.setCurrentPathway(null);
-      canvasState.getPathwayMap.delete(currentPathway.getUuid());
+      const posIndex = props.layer.pathways.findIndex((path) => path.isSame(currentPathway));
+      if (posIndex > -1) {
+        props.layer.pathways.splice(posIndex, 1);
+        currentPathway.destroy();
+        controller.setCurrentPathway(null);
+        canvasState.getPathwayMap.delete(currentPathway.getUuid());
+      }
     }
   });
 </script>
